# services/user_service/src/docs/swagger.yaml
definitions:
  main.APIError:
    properties:
      error:
        type: string
    type: object
  main.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  main.DefaultCreatedResponse:
    properties:
      id:
        type: string
    type: object
  main.LoginEmailRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  main.LoginResponse:
    properties:
      access_token:
        type: string
    type: object
  main.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/auth/login-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.LoginEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.APIError'
      summary: Login with email and password
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.DefaultCreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.APIError'
      summary: Register a new user
      tags:
      - auth
  /api/v1/health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /api/v1/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.User'
            type: array
      summary: Get all users
      tags:
      - users
  /api/v1/users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.APIError'
      summary: Get current user
      tags:
      - auth
swagger: "2.0"